+++ lfs.c
@@ -214,7 +214,7 @@ static int get_dir (lua_State *L) {
         path = realloc(path, size);
         if (!path) /* failed to allocate */
             return pusherror(L, "get_dir realloc() failed");
-        if (getcwd(path, size) != NULL) {
+        if (getcwd(path, (int)size) != NULL) {
             /* success, push the path to the Lua stack */
             lua_pushstring(L, path);
             result = 1;
